IMR-169 refactor: change db import path
